--- easynav_amcl/easynav_amcl_plugins.xml
@@ -1,9 +1,0 @@
-<class_libraries>
-  <library path="easynav_amcl">
-    <class name="easynav_amcl/AmclMethod"  type="easynav::AmclMethod" base_class_type="easynav::LocalizerMethodBase">
-      <description>
-      Nav2 AMCL implementation for EasyNavigation.
-      </description>
-    </class>
-  </library>
-</class_libraries>
