--- easynav_costmap_maps_manager/easynav_costmap_maps_manager_plugins.xml
@@ -1,9 +1,0 @@
-<class_libraries>
-  <library path="easynav_costmap_maps_manager">
-    <class name="easynav_costmap_maps_manager/CostmapMapsManager"  type="easynav::CostmapMapsManager" base_class_type="easynav::MapsManagerBase">
-      <description>
-      Costmap implementation for the Maps Manager.
-      </description>
-    </class>
-  </library>
-</class_libraries>
