--- easynav_costmap_planner/easynav_costmap_planner_plugins.xml
@@ -1,9 +1,0 @@
-<class_libraries>
-  <library path="easynav_costmap_planner">
-    <class name="easynav_costmap_planner/CostmapPlanner"  type="easynav::CostmapPlanner" base_class_type="easynav::PlannerMethodBase">
-      <description>
-      Easy Navigation Costmap Planner.
-      </description>
-    </class>
-  </library>
-</class_libraries>
