--- easynav_costmap_planner/package.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>easynav_costmap_planner</name>
-  <version>0.0.1</version>
-  <description>Easy Navigation: Costmap Planner package.</description>
-  <maintainer email="jc.manzanares.serrano@gmail.com">Juan Carlos Manzanares Serrano</maintainer>
-  <license>GPL-3.0-only</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <depend>rclcpp</depend>
-  <depend>rclcpp_lifecycle</depend>
-  <depend>easynav_common</depend>
-  <depend>easynav_core</depend>
-  <depend>pluginlib</depend>
-  <depend>nav2_planner</depend>
-  <depend>nav2_costmap_2d</depend>
-  <depend>easynav_costmap_maps_manager</depend>
-
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>ament_lint_common</test_depend>
-  <test_depend>ament_cmake_gtest</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
