+++ easynav_nav2_amcl/easynav_nav2_amcl_plugins.xml
@@ -1,0 +1,9 @@
+<class_libraries>
+  <library path="easynav_nav2_amcl">
+    <class name="easynav_nav2_amcl/AMCLLocalizer"  type="easynav::AMCLLocalizer" base_class_type="easynav::LocalizerMethodBase">
+      <description>
+      Nav2 AMCL implementation for EasyNavigation.
+      </description>
+    </class>
+  </library>
+</class_libraries>
