+++ easynav_nav2_planner/easynav_nav2_planner_plugins.xml
@@ -1,0 +1,9 @@
+<class_libraries>
+  <library path="easynav_nav2_planner">
+    <class name="easynav_nav2_planner/CostmapPlanner"  type="easynav::CostmapPlanner" base_class_type="easynav::PlannerMethodBase">
+      <description>
+      Easy Navigation Nav2 Costmap Planner.
+      </description>
+    </class>
+  </library>
+</class_libraries>
